test(InputTest): InputTest 분리
InputTest -> InputParserTest, InputValidatorTest 분리
